requirements.txt
certifi==2020.12.5
chardet==3.0.4
darkskylib==0.3.91
Django==2.2.5
django-extensions==3.1.1
djangorestframework==3.12.2
future==0.18.2
idna==2.10
pytz==2021.1
requests==2.23.0
sqlparse==0.4.1
urllib3==1.25.11
